+++ backend/docker-compose.dev.yml
@@ -12,6 +12,4 @@ services:
       - ./:/app
     ports:
       - "80:80"
-    environment:
-      - APP_ENV=development
     command: sh -c "uvicorn main:app --host 0.0.0.0 --port 80 --reload"
