+++ backend/requirements.txt
@@ -1,5 +1,8 @@
 fastapi[standard]==0.115.4
 pandas==2.2.3
+pydantic==2.9.2
+mlflow==2.8.0
+scikit-learn==1.3.0
 SQLAlchemy==2.0.36
 PyMySQL==1.1.1
 python-dotenv==1.0.1
