--- backend/src/controller/__init__.py
@@ -1,1 +1,0 @@
-from .hotel_controller import HotelController
