--- backend/src/decorator/__init__.py
@@ -1,1 +1,0 @@
-from .fastapi_decorator import controller, get, post
