--- backend/src/repository/__init__.py
@@ -1,1 +1,0 @@
-from .hotel_repository import HotelRepository
