--- backend/src/router/__init__.py
@@ -1,1 +1,0 @@
-from .router import router
