--- backend/src/service/__init__.py
@@ -1,1 +1,0 @@
-from .hotel_service import HotelService
