--- backend/src/utils/__init__.py
@@ -1,1 +1,0 @@
-from .db_connector import DBConnector
