+++ mlflow/requirements.txt
@@ -2,6 +2,6 @@ fastapi>=0.104.0
 uvicorn>=0.24.0
 pydantic>=2.0.0
 pandas>=2.0.0
-mlflow>=2.8.0
-scikit-learn>=1.3.0
+mlflow==2.8.0
+scikit-learn==1.3.0
 python-dotenv>=1.0.0
